feat: forge fmt

+++ test/QuietMode.t.sol
@@ -11,102 +11,119 @@ contract QuietModeTest is Test {
     using Senders for Senders.Sender;
     using Deployer for Senders.Sender;
     using Deployer for Deployer.Deployment;
-    
+
     TestScript normalScript;
     TestScript quietScript;
-    
+
     address constant TEST_ACCOUNT = 0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266;
     uint256 constant TEST_PRIVATE_KEY = 0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80;
-    
+
     function setUp() public {
         // Create normal script (not quiet)
         normalScript = new TestScript(false);
-        
+
         // Create quiet script
         quietScript = new TestScript(true);
-        
+
         // Fund test account
         vm.deal(TEST_ACCOUNT, 10 ether);
     }
-    
+
     function test_NormalModeEmitsEvents() public {
         // Record logs to check events
         vm.recordLogs();
-        
+
         // Execute transaction in normal mode
         normalScript.executeTransaction();
-        
+
         // Get logs
         Vm.Log[] memory logs = vm.getRecordedLogs();
-        
+
         // Should have TransactionSimulated event
         bool foundTransactionSimulated = false;
-        for (uint i = 0; i < logs.length; i++) {
-            if (logs[i].topics[0] == keccak256("TransactionSimulated(bytes32,address,address,uint256,bytes,string,bytes)")) {
+        for (uint256 i = 0; i < logs.length; i++) {
+            if (
+                logs[i].topics[0]
+                    == keccak256("TransactionSimulated(bytes32,address,address,uint256,bytes,string,bytes)")
+            ) {
                 foundTransactionSimulated = true;
                 break;
             }
         }
         assertTrue(foundTransactionSimulated, "TransactionSimulated event should be emitted in normal mode");
     }
-    
+
     function test_QuietModeSuppressesEvents() public {
         // Record logs to check events
         vm.recordLogs();
-        
+
         // Execute transaction in quiet mode
         quietScript.executeTransaction();
-        
+
         // Get logs
         Vm.Log[] memory logs = vm.getRecordedLogs();
-        
+
         // Should NOT have TransactionSimulated event
         bool foundTransactionSimulated = false;
-        for (uint i = 0; i < logs.length; i++) {
-            if (logs[i].topics[0] == keccak256("TransactionSimulated(bytes32,address,address,uint256,bytes,string,bytes)")) {
+        for (uint256 i = 0; i < logs.length; i++) {
+            if (
+                logs[i].topics[0]
+                    == keccak256("TransactionSimulated(bytes32,address,address,uint256,bytes,string,bytes)")
+            ) {
                 foundTransactionSimulated = true;
                 break;
             }
         }
         assertFalse(foundTransactionSimulated, "TransactionSimulated event should NOT be emitted in quiet mode");
     }
-    
+
     function test_QuietModeDeploymentDoesNotEmitEvents() public {
         // Record logs to check events
         vm.recordLogs();
-        
+
         // Deploy contract in quiet mode
         quietScript.deployContract();
-        
+
         // Get logs
         Vm.Log[] memory logs = vm.getRecordedLogs();
-        
+
         // Should NOT have ContractDeployed event
         bool foundContractDeployed = false;
-        for (uint i = 0; i < logs.length; i++) {
-            if (logs[i].topics[0] == keccak256("ContractDeployed(address,address,bytes32,((string,string,string,bytes32,bytes32,bytes32,bytes,string)))")) {
+        for (uint256 i = 0; i < logs.length; i++) {
+            if (
+                logs[i].topics[0]
+                    == keccak256(
+                        "ContractDeployed(address,address,bytes32,((string,string,string,bytes32,bytes32,bytes32,bytes,string)))"
+                    )
+            ) {
                 foundContractDeployed = true;
                 break;
             }
         }
         assertFalse(foundContractDeployed, "ContractDeployed event should NOT be emitted in quiet mode");
     }
-    
+
     function test_NormalModeDeploymentEmitsEvents() public {
         // Record logs to check events
         vm.recordLogs();
-        
+
         // Deploy contract in normal mode
         normalScript.deployContract();
-        
+
         // Get logs
         Vm.Log[] memory logs = vm.getRecordedLogs();
-        
+
         // Should have ContractDeployed event (but we'll also accept TransactionSimulated)
         bool foundEvent = false;
-        for (uint i = 0; i < logs.length; i++) {
-            if (logs[i].topics[0] == keccak256("ContractDeployed(address,address,bytes32,((string,string,string,bytes32,bytes32,bytes32,bytes,string)))") ||
-                logs[i].topics[0] == keccak256("TransactionSimulated(bytes32,address,address,uint256,bytes,string,bytes)")) {
+        for (uint256 i = 0; i < logs.length; i++) {
+            if (
+                logs[i].topics[0]
+                    == keccak256(
+                        "ContractDeployed(address,address,bytes32,((string,string,string,bytes32,bytes32,bytes32,bytes,string)))"
+                    )
+                    || logs[i].topics[0]
+                        == keccak256("TransactionSimulated(bytes32,address,address,uint256,bytes,string,bytes)")
+            ) {
                 foundEvent = true;
                 break;
             }
@@ -118,7 +135,7 @@ contract QuietModeTest is Test {
 // Simple test contract
 contract SimpleContract {
     uint256 public value;
-    
+
     function setValue(uint256 _value) public {
         value = _value;
     }
@@ -129,15 +146,17 @@ contract TestScript is ConfigurableTrebScript {
     using Senders for Senders.Sender;
     using Deployer for Senders.Sender;
     using Deployer for Deployer.Deployment;
-    
-    constructor(bool _quiet) ConfigurableTrebScript(
-        _getSenderConfigs(),
-        "test",
-        ".test-registry.json",
-        false,  // not dryrun
-        _quiet  // quiet mode
-    ) {}
-    
+
+    constructor(bool _quiet)
+        ConfigurableTrebScript(
+            _getSenderConfigs(),
+            "test",
+            ".test-registry.json",
+            false, // not dryrun
+            _quiet // quiet mode
+        )
+    {}
+
     function _getSenderConfigs() internal pure returns (Senders.SenderInitConfig[] memory) {
         Senders.SenderInitConfig[] memory configs = new Senders.SenderInitConfig[](1);
         configs[0] = Senders.SenderInitConfig({
@@ -148,13 +167,13 @@ contract TestScript is ConfigurableTrebScript {
         });
         return configs;
     }
-    
+
     function executeTransaction() public broadcast {
         Senders.Sender storage s = sender("test");
-        
+
         // Deploy a simple contract
         SimpleContract sc = new SimpleContract();
-        
+
         // Execute a transaction
         Transaction memory txn = Transaction({
             to: address(sc),
@@ -162,10 +181,10 @@ contract TestScript is ConfigurableTrebScript {
             value: 0,
             label: "setValue"
         });
-        
+
         s.execute(txn);
     }
-    
+
     function deployContract() public {
         // For this test, we'll just check that execute emits events
         // since deployment requires CreateX which isn't available in this test
